pyproject.toml
[tool.poetry]
name = "social-media-gen-llm-python"
version = "0.1.0"
description = "Social Media Posts Generator by AI"
authors = ["moosedev <ahmed.musawir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.15"
langchain = "^0.2.15"
langchain-core = "^0.2.37"
pydantic = "^2.8.2"
openai = "^1.43.0"
streamlit = "^1.38.0"
python-decouple = "^3.8"
bs4 = "^0.0.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
